requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.5
django-payments==1.0.0
django-phonenumber-field==6.1.0
idna==3.3
phonenumberslite==8.12.46
Pillow==9.1.0
psycopg2==2.9.3
pyTelegramBotAPI==4.4.1
python-decouple==3.6
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
